# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "adhocly-server",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local adhocly API"
    }
  ],
  "paths": {
    "/api/boards": {
      "get": {
        "tags": [
          "kanban"
        ],
        "operationId": "list_boards",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Board"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "kanban"
        ],
        "operationId": "create_board",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/boards/{id}": {
      "put": {
        "tags": [
          "kanban"
        ],
        "operationId": "update_board",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BoardInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Board"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "kanban"
        ],
        "operationId": "delete_board",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/contacts/{id}": {
      "put": {
        "tags": [
          "contexts"
        ],
        "operationId": "save_contact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contact"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "contexts"
        ],
        "operationId": "delete_contact",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
    "/api/context-links/{id}": {
      "put": {
        "tags": [
          "contexts"
        ],
        "operationId": "save_links",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextLinks"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextLinks"
                }
              }
            }
          }
        }
      }
    },
    "/api/contexts/{id}": {
      "put": {
        "tags": [
          "contexts"
        ],
        "operationId": "save_context",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkContext"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkContext"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "contexts"
        ],
        "operationId": "delete_context",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
    "/api/sync": {
      "get": {
        "tags": [
          "sync"
        ],
        "operationId": "getSyncSnapshot",
        "responses": {
          "200": {
            "description": "Atomic server snapshot and entity revisions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncSnapshot"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sync"
        ],
        "operationId": "applySyncOperation",
        "parameters": [
          {
            "name": "snapshot",
            "in": "query",
            "description": "Set false to omit the full snapshot from the reply; absent means true",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncOperation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Committed or previously committed mutation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncReply"
                }
              }
            }
          },
          "400": {
            "description": "Invalid mutation",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Revision or mutation id conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/task-fields/{id}": {
      "put": {
        "tags": [
          "kanban"
        ],
        "operationId": "update_field",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskField"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskField"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BTreeMap": {
        "type": "object",
        "additionalProperties": {
          "type": "integer",
          "format": "int64"
        },
        "propertyNames": {
          "type": "string"
        }
      },
      "Board": {
        "type": "object",
        "required": [
          "id",
          "name",
          "fieldId"
        ],
        "properties": {
          "fieldId": {
            "type": "string",
            "description": "Must be \"status\"; custom properties belong to contexts."
          },
          "id": {
            "type": "string"
          },
          "manualOrder": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "sortMode": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "BoardInput": {
        "type": "object",
        "required": [
          "id",
          "name",
          "fieldId"
        ],
        "properties": {
          "fieldId": {
            "type": "string",
            "description": "Must be \"status\"; custom properties belong to contexts."
          },
          "id": {
            "type": "string"
          },
          "manualOrder": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "name": {
            "type": "string"
          },
          "sortMode": {
            "type": [
              "string",
              "null"
            ],
            "description": "Missing sorting fields preserve existing settings for older clients."
          }
        },
        "additionalProperties": false
      },
      "Contact": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ContextField": {
        "type": "object",
        "required": [
          "id",
          "name",
          "kind"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ContextFieldKind"
          },
          "name": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "value": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ContextFieldKind": {
        "type": "string",
        "enum": [
          "identifier",
          "contact",
          "choice",
          "text",
          "date",
          "link",
          "note"
        ]
      },
      "ContextLinks": {
        "type": "object",
        "description": "Each owner has one independently revised attachment record. Empty values mask inheritance;\nabsent keys inherit. Field ids are scoped to a context, so unrelated schemas never collide.",
        "required": [
          "id",
          "contextIds",
          "overrides"
        ],
        "properties": {
          "contextIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "id": {
            "type": "string"
          },
          "overrides": {
            "type": "object",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              },
              "propertyNames": {
                "type": "string"
              }
            },
            "propertyNames": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "CustomReminder": {
        "type": "object",
        "required": [
          "id",
          "kind"
        ],
        "properties": {
          "at": {
            "type": [
              "string",
              "null"
            ],
            "description": "Local wall-clock time; required only for custom-date reminders."
          },
          "id": {
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ReminderKind"
          },
          "offsetMinutes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Legacy presets; omit when using offsetUnit and offsetValue."
          },
          "offsetUnit": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReminderUnit"
              }
            ]
          },
          "offsetValue": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Whole units before the anchor, from 1 through 999."
          },
          "urgent": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "FieldOption": {
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "Folder": {
        "type": "object",
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "parentId": {
            "type": [
              "string",
              "null"
            ],
            "description": "`None` is the top level; folders nest to any depth."
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "Project": {
        "type": "object",
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "folderId": {
            "type": [
              "string",
              "null"
            ],
            "description": "`None` means the project sits at the top level of the sidebar."
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "ReminderKind": {
        "type": "string",
        "enum": [
          "custom",
          "planned",
          "due"
        ]
      },
...
      "SyncOperation": {
        "type": "object",
        "required": [
          "id",
          "url",
          "method",
          "body",
          "expectedVersion"
        ],
        "properties": {
          "body": {},
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "SyncReply": {
        "type": "object",
        "required": [
          "changes"
        ],
        "properties": {
          "changes": {
            "$ref": "#/components/schemas/BTreeMap",
            "description": "Only versions changed by this operation, not unrelated changes from another device."
          },
          "snapshot": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SyncSnapshot",
                "description": "Omitted when the caller asked for `?snapshot=false`: the client only needs the snapshot for\nthe last operation of a batch, and a whole-dataset reply per operation is the dominant cost."
              }
            ]
          }
        }
      },
      "SyncSnapshot": {
        "type": "object",
        "required": [
          "protocolVersion",
          "contexts",
          "contacts",
          "contextLinks",
          "tasks",
          "projects",
          "folders",
          "taskFields",
          "boards",
          "versions"
        ],
        "properties": {
          "boards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Board"
            }
          },
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contact"
            }
          },
          "contextLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContextLinks"
            }
          },
          "contexts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkContext"
            }
          },
          "folders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Folder"
            }
          },
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          },
          "protocolVersion": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "taskFields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TaskField"
            }
          },
          "tasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          },
          "versions": {
            "$ref": "#/components/schemas/BTreeMap"
          }
        }
      },
      "Task": {
        "type": "object",
        "required": [
          "id",
          "title",
          "projectId",
          "project",
          "createdAt",
          "updatedAt",
          "completed"
        ],
        "properties": {
          "completed": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "dueOn": {
            "type": [
              "string",
              "null"
            ],
            "description": "Local wall-clock date and time, `YYYY-MM-DDTHH:MM`."
          },
          "id": {
            "type": "string"
          },
          "plannedFor": {
            "type": [
              "string",
              "null"
            ],
            "description": "Local wall-clock date and time, `YYYY-MM-DDTHH:MM`."
          },
          "project": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "reminders": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomReminder"
            }
          },
          "repeatWeekday": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "statusId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "TaskField": {
        "type": "object",
        "description": "Shared task status configuration; the historical wire shape is retained for sync retries.",
        "required": [
          "id",
          "name",
          "kind",
          "options"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Must be \"status\"."
          },
          "kind": {
            "type": "string",
            "description": "Must be \"choice\"."
          },
          "name": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldOption"
            }
          }
        },
        "additionalProperties": false
      },
...
      "WorkContext": {
        "type": "object",
        "required": [
          "id",
          "name",
          "fields"
        ],
        "properties": {
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContextField"
            }
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      }
    }
  },
...
}